+++ components/debug-views/react-controls.ts
@@ -1,0 +1,1 @@
+export { useDebugViewsControls } from "./use-debug-views-controls"