+++ vite.config.ts
@@ -18,6 +18,7 @@ export default defineConfig({
       entry: {
         index: resolve(__dirname, "components/debug-views/index.ts"),
         react: resolve(__dirname, "components/debug-views/react.ts"),
+        "react-controls": resolve(__dirname, "components/debug-views/react-controls.ts"),
       },
       formats: ["es"],
     },